+++ constant/os.go
@@ -1,7 +1,6 @@
 package constant
 
 const (
-	Windows = "windows"
 	Darwin  = "darwin"
 	Linux   = "linux"
 	Android = "android"