+++ open/open.go
@@ -3,22 +3,16 @@ package open
 import (
 	"fmt"
 	"github.com/yukiteruamano/koma/constant"
-	"os"
 	"os/exec"
-	"path/filepath"
 	"runtime"
-	"strings"
 )
 
 var (
 	errUnsupportedOS = fmt.Errorf("can't open on this OS: %s", runtime.GOOS)
-	runDll32         = filepath.Join(os.Getenv("SYSTEMROOT"), "System32", "rundll32.exe")
 )
 
 func open(input string) (cmd *exec.Cmd, osSupported bool) {
 	switch runtime.GOOS {
-	case constant.Windows:
-		return exec.Command(runDll32, "url.dll,FileProtocolHandler", input), true
 	case constant.Darwin:
 		return exec.Command("open", input), true
 	case constant.Linux:
@@ -32,8 +26,6 @@ func open(input string) (cmd *exec.Cmd, osSupported bool) {
 
 func openWith(input, with string) (cmd *exec.Cmd, osSupported bool) {
 	switch runtime.GOOS {
-	case constant.Windows:
-		return exec.Command("cmd", "/C", "start", "", with, strings.ReplaceAll(input, "&", "^&")), true
 	case constant.Darwin:
 		return exec.Command("open", "-a", with, input), true
 	case constant.Linux: